Moved utility functions from calculator to utils
